refatorando o projeto

+++ src/main/java/br/com/toto/repository/Jogadores.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.toto.model.Jogador;
 
-public interface Jogadores extends JpaRepository<Jogador, Long>{
+public interface Jogadores extends JpaRepository<Jogador, Integer>{
 
 }
